🚚 Rename TrackersManager to TrackerManager
