+++ src/components/TrackerManager.vue
